feat: take the secondary index values from the query bounds
