+++ parsil/src/lib.rs
@@ -7,6 +7,7 @@ pub use utils::DEFAULT_MAX_BLOCK_PLACEHOLDER;
 pub use utils::DEFAULT_MIN_BLOCK_PLACEHOLDER;
 
 pub mod assembler;
+pub mod bracketer;
 pub mod errors;
 pub mod executor;
 mod expand;
